Make sure to enable input channel binding for websocket sink
 - Add `@EnableBinding` in the configuration class

This resolves #28
